Add nominal "output" port too.
It exists because such MIDI 2.0 devices that conform to the latest UMP
specification would like to connect to both input and output as a valid
by-directional pair of connections, and send UMP Stream Configuration
Notification as a result of *our* UMP Stream Configuration Request message
when we turn on MIDI2 mode.

Here, we prepare such a slot (MIDI OUT port, or MIDI "input" port in Android
MIDI API wording) so that clients could send the replies, which we ignore
regardless.