It seems there is no way to use compileSdk conditionally, so disable VIC build.
While `productFlavors { create(...) { compileSdkPreview = ... }}` looks like
valid, it actually replaces the whole compileSdkPreview specification for
the entire build, as `compileSdk(Preview)`, unlike `targetSdk` or `minSdk` is
NOT part of `AppliationProductFlavor`.
https://developer.android.com/reference/tools/gradle-api/8.3/com/android/build/api/dsl/ApplicationProductFlavor

As a result, as the build settings indicate that `compileSdkPreview` = VIC is
used even for `stable` flavor, and results in Play Store rejection.

It looks like an overlook by AGP and we cannot really have a valid setup for
VIC build so far. We will have to occasionally disable VIC build like this time.